backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Vector database and embeddings
chromadb==0.4.24
sentence-transformers==5.0.0

# Agent orchestration
langchain==0.2.16
langchain-core==0.2.40
langchain-openai==0.1.25
langgraph==0.2.28

# Document processing
pdfplumber==0.10.0
python-docx==1.1.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.27.0

# Optional: Production deployment
gunicorn==21.2.0

# Compatibility (for some deployment environments)
# pysqlite3-binary==0.5.2.post2
